changes in XssfReadder signature methods

+++ src/main/java/com/portal/client/service/microsoft/excel/reader/XssfReader.java
@@ -8,7 +8,7 @@ import com.portal.client.service.microsoft.excel.RowObject;
 
 public interface XssfReader {
 
-	List<RowObject> read(InputStream xlsxInputStream, short initialOffset, short endOffset) throws IOException;
+	List<RowObject> read(InputStream xlsxInputStream, int initialOffset, int endOffset) throws IOException;
 
-	List<RowObject> read(byte[] xlsxStreams, short initialOffset, short endOffset) throws IOException;
+	List<RowObject> read(byte[] xlsxStreams, int initialOffset, int endOffset) throws IOException;
 }
